arm: Add VMEXIT test skeleton
Add a VMEXIT test skeleton to the unit-tests.  This test is not that
easy to execute thought, because QEMU doesn't emulate a proper read of
the cycle counter and KVM guests cannot read the cycle counter on ARM.

You can enable guest access to the debug registers on KVM with a hack
to your host kernel, but it is dangerous and only advisable for the
truly courageous souls.

Still, this provides a useful framework to have in this package and will
only require minor modifications once we come up with a good solution
for VMs to read the cycle counter and establish a HVC number reserved
for tests like these.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>