arm: Add smp support and a simple smptest
Add (limited) functionality to boot a secondary CPU using PSCI, which is
necessary to measure things such as IPI overhead.  Things get more
complicted when we start enabling the MMU and such, but for now this
will do.

We do not support a general inderection for booting CPUs but rely on
callers to call the PSCI code directly.  We also copy in the KVM
definitions for the PSCI v0.1 function IDs, which are exported to user
space.

Some day we should do like QEMU and keep a local imported copy of the
KVM header files.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>