arm: SMP thread_info support
Support a small per-cpu thread-info structure located at the bottom of
each stack page for each CPU.  This struct contains only the cpu_id for
now, but allows any code to call get_cpu_id() to get the current CPU id
of the exeuting processor.

Also expand the smptest framework to test that CPU0 and CPU1 report this
value as expected.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>