arm: Add lib/arm/compiler.h
This is a simple file to add to built-in compiler options useful for all
sorts of test code.  We should have a structure where we had an
arch-generic compiler.h file which could automatically import an
arch-specific one, and the generic one could hold defines for gcc stuff
like 'noinline' and the arch-specific stuff could contain support for
stuff like arch_extensions, dsb etc. for ARM, for example.

Do with the ARM one for now.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>