arm: Fix and synchronize ARM puts implementation
Use writeb to output bytes to the pl011 serial port instead of pointer
dereference.

Add a spinlock around the puts operations, except if the MMU is
disabled, since the spin_lock() operation will never return in that case
(strex fails to update memory).  Allowing a print without taking the
spinlock when the MMU is disabled does not interfere with the validity
of testing and allows early boot code (like the MMU enablement code
itslef) to do early printf or debug prints, while synchronizing
remaining SMP code.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>