arm: Restructure exception handling
Restructure exception handling to allow each separate core to install a
separate exception handler.  This is useful for SMP tests where an IRQ
may be a catastrophic failure on one core, but a valid thing on another
core and at the same time you are measuring latency for the IRQ handlers
with as little conditional code in there as possible.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>