arm: Fix get_cpu_thread_info() bug and type warning
get_cpu_thread_info() will find the lower region of your stack page, but
if the stack pointer should happen to be at the very top of the stack
page, just masking the lower bits will return the wrong page.  OK, this
is actually impossible once we're in a C function and some state has to
be spilled on the stack, but this is called from early assembly code and
we are not sure which compiler optimizations would cause this to
happen.

Also do an explicit cast of sp to (void *)sp in get_cpu_id() to satisfy
GCC.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>