arm: vmexit: Let individual tests to cycle reads
Since we changed the IPI test to wait until the IPI receiving CPU is
ready, there are cycles for each run of the test that should not be
counted in the overall test, therefore let the test implementation
itself do the measurements and return the value to the caller.  This
requires some mangling around, but the result is more flexible.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>